pyproject.toml
[tool.poetry]
name = "pyart"
version = "0.1.0"
description = "A turtle graphics framework using pycairo"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pycairo = "^1.24.0"
colorspace = "^1.0.0"
matplotlib = "^3.10.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
include = ["pyart", "example.py"]
exclude = ["**/__pycache__", "**/node_modules"]
pythonVersion = "3.8"
pythonPlatform = "Linux"
typeCheckingMode = "basic"
reportMissingImports = true
reportMissingTypeStubs = false

